+++ NEWS.md
@@ -7,6 +7,19 @@ Note that each entry is kept to a minimum, see links for details.
 
 ## Language changes
 
+* Anonymous rest and keyword rest arguments can now be passed as
+  arguments, instead of just used in method parameters.
+  [[Feature #18351]]
+
+    ```ruby
+    def foo(*)
+      bar(*)
+    end
+    def baz(**)
+      quux(**)
+    end
+    ```
+
 ## Command line options
 
 ## Core classes updates
@@ -52,3 +65,5 @@ Note: Excluding feature bug fixes.
 ## IRB Autocomplete and Document Display
 
 ## Miscellaneous changes
+
+[Feature #18351]: https://bugs.ruby-lang.org/issues/18351